+++ src/main/java/saphy/saphy/coupon/domain/Coupon.java
@@ -5,8 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import saphy.saphy.global.BaseEntity;
-import saphy.saphy.item.domain.Item;
+import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.user.domain.User;
 
 @Entity
