+++ src/main/java/saphy/saphy/delivery/domain/Delivery.java
@@ -5,10 +5,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import saphy.saphy.global.BaseEntity;
-import saphy.saphy.item.domain.Item;
+import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.order.domain.Order;
-import saphy.saphy.user.domain.User;
 
 import java.time.LocalDateTime;
