+++ src/main/java/saphy/saphy/repairCenterBookmark/domain/RepairCenterBookmark.java
@@ -5,9 +5,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import saphy.saphy.global.BaseEntity;
+import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.repairCenter.domain.RepairCenter;
-import saphy.saphy.review.domain.Review;
 import saphy.saphy.user.domain.User;
 
 @Entity
