+++ src/main/java/saphy/saphy/repairCenterReview/domain/RepairCenterReview.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import saphy.saphy.global.BaseEntity;
+import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.repairCenter.domain.RepairCenter;
 import saphy.saphy.user.domain.User;
